requirements.txt
numpy==1.14.2
matplotlib==2.2.0
setuptools==39.0.1
untangle==1.1.1
pytest==3.4.2
pytest-cov==2.5.1
python-evtx==0.6.1